refactor: move Truncate to Utils dir
