+++ app/components/UI/index.js
@@ -38,6 +38,5 @@ export Text from './Text'
 export TextArea from './TextArea'
 export Titlebar from './Titlebar'
 export Toggle from './Toggle'
-export Truncate from './Truncate'
 export Value from './Value'
 export Wizard from './Wizard'
