+++ app/components/Util/index.js
@@ -1,1 +1,2 @@
+export Truncate from './Truncate'
 export WalletName from './WalletName'
